feat(sdk): Move sdk.py to utils
