+++ app/__init__.py
@@ -1,5 +1,5 @@
 """Agentic AI for Data Pipeline Incident Resolution Demo."""
 
-from app.entrypoints.sdk import run_investigation
+from app.utils.sdk import run_investigation
 
 __all__ = ["run_investigation"]
